+++ application/recipe/dto/recipe_ingredient_dtos.py
@@ -9,11 +9,11 @@ class RecipeIngredientCreateDto(BaseRecipeIngredientDto):
     quantity: int
     ingredient_id: int
     measurement_unit_id: int
-    recipe_id: int
-    
+
 
 class RecipeIngredientDto(RecipeIngredientCreateDto):
     id: int
+    recipe_id: int
     
 
 class RecipeIngredientUpdateDto(BaseRecipeIngredientDto):
