+++ application/recipe/services/recipe_service.py
@@ -110,41 +110,61 @@ class RecipeService:
             raise application_exceptions.EntityNotFoundApplicationException()
         return self._recipe_category_repository.delete(id)
 
-    def create_recipe_ingredient(self, recipe_ingredient_create_dto: RecipeIngredientCreateDto) -> RecipeIngredientDto:
-        ingredient = RecipeIngredient(id=None, quantity=recipe_ingredient_create_dto.quantity, ingredient_id=recipe_ingredient_create_dto.ingredient_id, ingredient_measurement_unit_id=recipe_ingredient_create_dto.measurement_unit_id, recipe_id=recipe_ingredient_create_dto.recipe_id)
+    def create_recipe_ingredient(self, recipe_id: int, recipe_ingredient_create_dto: RecipeIngredientCreateDto) -> RecipeIngredientDto:
         try:
-            created_ingredient = self._recipe_ingredient_repository.save(ingredient)
+            recipe = self._recipe_repository.get_by_id(recipe_id)
+        except domain_exceptions.NotFoundDomainException:
+            raise application_exceptions.EntityNotFoundApplicationException()
+        ingredient = RecipeIngredient(id=None, quantity=recipe_ingredient_create_dto.quantity, ingredient_id=recipe_ingredient_create_dto.ingredient_id, ingredient_measurement_unit_id=recipe_ingredient_create_dto.measurement_unit_id, recipe_id=recipe_id)
+        try:
+            created_ingredient = self._recipe_ingredient_repository.save(recipe_id=recipe_id, ingredient=ingredient)
         except domain_exceptions.DatabaseIntegrityDomainException:
             raise application_exceptions.EntityValidationApplicationException()
         return RecipeIngredientDto(id=created_ingredient.id, quantity=created_ingredient.quantity, ingredient_id=created_ingredient.ingredient_id, measurement_unit_id=created_ingredient.ingredient_measurement_unit_id, recipe_id=created_ingredient.recipe_id)
 
-    def get_recipe_ingredient(self, id) -> RecipeIngredientDto:
+    def get_recipe_ingredient(self, recipe_id: int, id: int) -> RecipeIngredientDto:
         try:
-            ingredient = self._recipe_ingredient_repository.get_by_id(id)
+            recipe = self._recipe_repository.get_by_id(recipe_id)
+        except domain_exceptions.NotFoundDomainException:
+            raise application_exceptions.EntityNotFoundApplicationException()
+        try:
+            ingredient = self._recipe_ingredient_repository.get_by_id(recipe_id=recipe_id, id=id)
         except domain_exceptions.NotFoundDomainException:
             raise application_exceptions.EntityNotFoundApplicationException
         return RecipeIngredientDto(id=ingredient.id, quantity=ingredient.quantity, ingredient_id=ingredient.ingredient_id, measurement_unit_id=ingredient.ingredient_measurement_unit_id, recipe_id=ingredient.recipe_id)
 
-    def list_recipe_ingredients(self, limit: int = 1000, offset: int = 0) -> List[RecipeIngredientDto]:
-        ingredients = self._recipe_ingredient_repository.get_all(limit=limit, offset=offset)
+    def list_recipe_ingredients(self, recipe_id: int, limit: int = 1000, offset: int = 0) -> List[RecipeIngredientDto]:
+        try:
+            recipe = self._recipe_repository.get_by_id(recipe_id)
+        except domain_exceptions.NotFoundDomainException:
+            raise application_exceptions.EntityNotFoundApplicationException()
+        ingredients = self._recipe_ingredient_repository.get_all(recipe_id=recipe_id, limit=limit, offset=offset)
         return [RecipeIngredientDto(id=i.id, quantity=i.quantity, ingredient_id=i.ingredient_id, recipe_id=i.recipe_id, measurement_unit_id=i.ingredient_measurement_unit_id) for i in ingredients]
 
-    def update_recipe_ingredient(self, id: int, ingredient_update_dto: RecipeIngredientUpdateDto) -> RecipeIngredientDto:
+    def update_recipe_ingredient(self, recipe_id: int, id: int, ingredient_update_dto: RecipeIngredientUpdateDto) -> RecipeIngredientDto:
         try:
-            ingredient = self._recipe_ingredient_repository.get_by_id(id)
+            recipe = self._recipe_repository.get_by_id(recipe_id)
+        except domain_exceptions.NotFoundDomainException:
+            raise application_exceptions.EntityNotFoundApplicationException()
+        try:
+            ingredient = self._recipe_ingredient_repository.get_by_id(recipe_id=recipe_id, id=id)
         except domain_exceptions.NotFoundDomainException:
             raise application_exceptions.EntityNotFoundApplicationException()
         ingredient.ingredient_measurement_unit_id = ingredient_update_dto.measurement_unit_id
         ingredient.quantity = ingredient_update_dto.quantity
         try:
-            updated_ingredient = self._recipe_ingredient_repository.save(ingredient)
+            updated_ingredient = self._recipe_ingredient_repository.save(recipe_id=recipe_id, ingredient=ingredient)
         except domain_exceptions.DatabaseIntegrityDomainException:
             raise application_exceptions.EntityValidationApplicationException()
         return RecipeIngredientDto(id=updated_ingredient.id, quantity=updated_ingredient.quantity, ingredient_id=updated_ingredient.ingredient_id, recipe_id=updated_ingredient.recipe_id, measurement_unit_id=updated_ingredient.ingredient_measurement_unit_id)
 
-    def delete_recipe_ingredient(self, id) -> None:
+    def delete_recipe_ingredient(self, recipe_id: int, id: int) -> None:
         try:
-            return self._recipe_ingredient_repository.delete(id)
+            recipe = self._recipe_repository.get_by_id(recipe_id)
+        except domain_exceptions.NotFoundDomainException:
+            raise application_exceptions.EntityNotFoundApplicationException()
+        try:
+            return self._recipe_ingredient_repository.delete(recipe_id=recipe_id, id=id)
         except domain_exceptions.NotFoundDomainException:
             raise application_exceptions.EntityNotFoundApplicationException()
