+++ infrastructure/persistence/sql_alchemy/repositories/ingredient_repository.py
@@ -54,5 +54,3 @@ class IngredientSqlAlchemyRepository(BaseSqlAlchemyRepository, IngredientReposit
                 return orm_object.to_domain()
             except IntegrityError as e:
                 raise domain_exceptions.DatabaseIntegrityDomainException()
-            except Exception as e:
-                pass
